Add initial QGMLWY + layers configuration
This makes the default layer QGMLWY: https://mk.bcgsc.ca/carpalx/?full_optimization. Adds QWERTY and Colemak layers as alternatives to those where only the letters are different (unusual placement of modifiers, symbols, etc still applies.

Adds a "symbols" layer that changes the home row to be the symbol-alternates from the normal 1-8 keys with -/_ and =/+ swapped out for 9 and 0 alternates. This may in the future become the number keys at home row with the alternates through shift for more ergonomic typing.

Adds a "control" layer that has macros specifically for controlling Rectable on macOS.

Adds a "media" layer that mimics the macOS media keys in as cross-platform way as possible (and then a macro and generic function key mapping to compensate when necessary).

Copies over the "mod" and "fn" default layers.